+++ app/src/main/java/com/machado/thenew20hourrule/presentation/skill_detail_screen/SkillDetailFragment.kt
@@ -153,7 +153,7 @@ class SkillDetailFragment : Fragment() {
 
                             val newSession = Session(
                                 objective = sessionObjective.text.toString(),
-                                sessionDurationInMin = sessionDuration.text.toString().toDouble(),
+                                sessionDurationInMin = sessionDuration.text.toString().toFloat(),
                                 skillId = skill.skillId!!
                             )
                             // Set datetime when done
